Introduce the 4pi project; make inital commit
4pi is a collection of simple computer programs that estimate the value
of Pi, the mathematical constant defined as the ratio of a circle's
circumference to its diameter. Each program in the collection differs
only in the programming language it was written in, the set of features
of the language it utilized, and/or the fundamental underlying
mathematical algorithm it implemented to approximate the value of Pi.

The principal aim of the 4pi project is to explore different aspects of
each programming language and their feature sets from a scientific and
high-performance computing perspective. For example, the first set of
programs included in this initial commit to the project estimate the
value of Pi via the Monte Carlo method. This solution is particularly
useful for exploring different parallel programming models, languages,
libraries, and APIs as it is an embarrassingly parallel (albeit
inefficient) solution to the problem.

The programming languagues represented in this initial commit are bash,
fortran, and python. A C program is also planned.

In the future, additional mathematical algorithms to approximate the
value of Pi --- beyond the Monte Carlo approach --- will also be
considered and implemented. These alternate mathematical approaches will
allow the project explore an even wider field of programming languages
and feature sets over time as it grows.